[fix] Print absolute value

+++ 0x02-functions_nested_loops/6-abs.c
@@ -5,7 +5,7 @@
  * Return: Absolute value of parameter
  */
 
-int _abs(int a)
+double _abs(double a)
 {
 	return (fabs(a));
 }
